backend/attendance_system/requirements.txt

# Django and REST framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-extensions==3.2.3

mysqlclient==2.2.0

Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.26.2
tensorflow==2.15.0  # For deep learning-based face recognition
scikit-learn==1.3.2  # For machine learning utilities
matplotlib==3.8.2    # For visualization

# Data handling
pandas==2.1.3
faker==20.1.0

# HTTP client for ESP32-CAM integration
requests==2.31.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
tqdm==4.66.1  # Progress bar for model training
